Moved simple interest calculator files to solutions and tests folder
