+++ solutions/simple_interest_calculator.py
